move things around

+++ mod/home/dot/.profile
@@ -5,22 +5,21 @@ export LC_ALL="en_US.UTF-8"
 
 case " ${ok:-false} $- " in
 (" false "*i*" ")
-	export PS1="\$USER@$(hostname -s):\$PWD%\${?#0} "
-	export SSH_AUTH_SOCK="$HOME/.ssh/sock/agent"
-	export LESSHISTFILE="/tmp/lesshist"
 	export EDITOR="vim"
-	export LESS="XSR"
 	export PAGER="less"
 	export MANPAGER="less"
+	export LESSHISTFILE="/tmp/lesshist"
+	export PS1="$(hostname -s):\$PWD\\\$\${?#0} "
 
-	type w >/dev/null && w
-
-	if [ -z "$SSH_CLIENT" ] &&
-	! pgrep -fx "ssh-agent -a $SSH_AUTH_SOCK" >/dev/null; then
-		rm -f "$SSH_AUTH_SOCK"
-		ssh-agent -a "$SSH_AUTH_SOCK"
+	if [ -z "$SSH_CONNECTION" ]; then
+		export SSH_AUTH_SOCK="$HOME/.ssh/sock/agent"
+		if ! pgrep -fx "ssh-agent -a $SSH_AUTH_SOCK" >/dev/null; then
+			rm -f "$SSH_AUTH_SOCK"
+			ssh-agent -a "$SSH_AUTH_SOCK"
+		fi
 	fi
 	;;
 esac
+
 set -o emacs 2>"/dev/null"
 ok=true
